Main Menu Added
Added a Main Menu scene which uses the stage as the background. Has the title up top and three buttons, a start button, a credits button and a quit button. The credits button changes the menu to show our names and roles and has a new back button which returns you to the main menu.